Get products info & set those data in UI

+++ E-Market Demo/E-Market Demo/Model/Order.swift
@@ -23,7 +23,7 @@ public struct Order: Codable {
 //public struct Product: DictionaryConvertor {
 public struct Product: Codable {
     let name: String?
-    let price: Int?
+    let price: Double?
     let imageURL: String?
 
     enum CodingKeys: String, CodingKey {
